+++ packages/stack-server/src/app/(main)/(protected)/projects/[projectId]/sidebar-layout.tsx
@@ -18,7 +18,7 @@ import { Link as LinkIcon } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { Button, buttonVariants } from "@/components/ui/button";
 import { Project, UserButton, useUser } from "@stackframe/stack";
-import { usePathname } from "next/navigation";
+import { usePathname, useRouter } from "next/navigation";
 import { Fragment, useMemo, useState } from "react";
 import {
   Breadcrumb,
@@ -190,7 +190,7 @@ function NavItem({ item, href, onClick }: { item: Item, href: string, onClick?: 
   const selected = useMemo(() => {
     return item.regex.test(pathname);
   }, [item.regex, pathname]);
-  
+
   return (
     <Link
       href={href}
@@ -200,6 +200,7 @@ function NavItem({ item, href, onClick }: { item: Item, href: string, onClick?: 
         "flex-grow justify-start text-sm text-zinc-800 dark:text-zinc-300 px-2",
       )}
       onClick={onClick}
+      prefetch={true}
     >
       <item.icon className="mr-2 h-4 w-4" />
       {item.name}