+++ packages/stack-server/src/app/api/v1/teams/route.tsx
@@ -7,6 +7,7 @@ import { checkApiKeySet, secretServerKeyHeaderSchema } from "@/lib/api-keys";
 import { isProjectAdmin } from "@/lib/projects";
 import { createServerTeam, listServerTeams } from "@/lib/teams";
 import { ServerTeamJson } from "@stackframe/stack-shared/dist/interface/serverInterface";
+import { KnownErrors } from "@stackframe/stack-shared";
 
 const getSchema = yup.object({
   query: yup.object({
@@ -37,7 +38,7 @@ export const GET = deprecatedSmartRouteHandler(async (req: NextRequest) => {
   let teams: ServerTeamJson[] = [];
   if (server === "true") {
     if (!skValid && !asValid) {
-      throw new StatusError(StatusError.Forbidden);
+      throw new KnownErrors.ApiKeyNotFound();
     }
     teams = await listServerTeams(projectId);
   }