+++ packages/stack/src/lib/stack-app.ts
@@ -9,7 +9,7 @@ import { AsyncStore } from "@stackframe/stack-shared/dist/utils/stores";
 import { ClientProjectJson, UserJson, TokenObject, TokenStore, ProjectJson, EmailConfigJson, DomainConfigJson, ReadonlyTokenStore, getProductionModeErrors, ProductionModeError, UserUpdateJson, TeamJson, PermissionDefinitionJson, PermissionDefinitionScopeJson, TeamMemberJson } from "@stackframe/stack-shared/dist/interface/clientInterface";
 import { isClient } from "../utils/next";
 import { callOAuthCallback, signInWithOAuth } from "./auth";
-import * as NextNavigation from "next/navigation";  // import the entire module to get around some static compiler warnings emitted by Next.js in some cases
+import * as NextNavigationUnscrambled from "next/navigation";  // import the entire module to get around some static compiler warnings emitted by Next.js in some cases
 import { ReadonlyJson } from "@stackframe/stack-shared/dist/utils/json";
 import { constructRedirectUrl } from "../utils/url";
 import { filterUndefined, omit } from "@stackframe/stack-shared/dist/utils/objects";
@@ -19,13 +19,24 @@ import { ApiKeySetBaseJson, ApiKeySetCreateOptions, ApiKeySetFirstViewJson, ApiK
 import { suspend } from "@stackframe/stack-shared/dist/utils/react";
 import { EmailTemplateType, ServerPermissionDefinitionCustomizableJson, ServerPermissionDefinitionJson, ServerTeamCustomizableJson, ServerTeamJson, ServerTeamMemberJson, ServerUserUpdateJson } from "@stackframe/stack-shared/dist/interface/serverInterface";
 import { EmailTemplateCrud, ListEmailTemplatesCrud } from "@stackframe/stack-shared/dist/interface/crud/email-templates";
+import { scrambleDuringCompileTime } from "@stackframe/stack-shared/dist/utils/compile-time";
+import { isReactServer } from "@stackframe/stack-sc";
+import * as cookie from "cookie";
+
+// NextNavigation.useRouter does not exist in react-server environments and some bundler try to be helpful and throw a warning. Ignore the warning.
+const NextNavigation = scrambleDuringCompileTime(NextNavigationUnscrambled);
 
 const clientVersion = process.env.STACK_COMPILE_TIME_CLIENT_PACKAGE_VERSION ?? throwErr("Missing STACK_COMPILE_TIME_CLIENT_PACKAGE_VERSION. This should be a compile-time variable set by Stack's build system.");
 
-export type TokenStoreOptions<HasTokenStore extends boolean = boolean> =
-  HasTokenStore extends true ? "cookie" | "nextjs-cookie" | "memory" :
-  HasTokenStore extends false ? null :
-  TokenStoreOptions<true> | TokenStoreOptions<false>;
+export type TokenStoreInit<HasTokenStore extends boolean = boolean> =
+  HasTokenStore extends true ? (
+    | "cookie"
+    | "nextjs-cookie"
+    | "memory"
+    | Request
+  )
+  : HasTokenStore extends false ? null
+  : TokenStoreInit<true> | TokenStoreInit<false>;
 
 export type HandlerUrls = {
   handler: string,
@@ -98,8 +109,8 @@ export type StackClientAppConstructorOptions<HasTokenStore extends boolean, Proj
   publishableClientKey?: string,
   urls?: Partial<HandlerUrls>,
 
-  // we intersect with TokenStoreOptions in the beginning to make TypeScript error messages easier to read
-  tokenStore: TokenStoreOptions<HasTokenStore>,
+  // we intersect with TokenStoreInit in the beginning to make TypeScript error messages easier to read
+  tokenStore: TokenStoreInit<HasTokenStore>,
 };
 
 export type StackServerAppConstructorOptions<HasTokenStore extends boolean, ProjectId extends string> = StackClientAppConstructorOptions<HasTokenStore, ProjectId> & {
@@ -116,7 +127,8 @@ export type StackAdminAppConstructorOptions<HasTokenStore extends boolean, Proje
   | (
     & Omit<StackServerAppConstructorOptions<HasTokenStore, ProjectId>, "publishableClientKey" | "secretServerKey">
     & {
-      projectOwnerTokens: ReadonlyTokenStore,
+      projectOwnerTokens: TokenStore,
+      refreshProjectOwnerTokens: () => Promise<void>,
     }
   )
 );
@@ -129,15 +141,12 @@ export type StackClientAppJson<HasTokenStore extends boolean, ProjectId extends 
 const defaultBaseUrl = "https://app.stack-auth.com";
 
 function createEmptyTokenStore() {
-  const store = new AsyncStore<TokenObject>();
-  store.set({
+  return new AsyncStore<TokenObject>({
     refreshToken: null,
     accessToken: null,
   });
-  return store;
 }
 
-const memoryTokenStore = createEmptyTokenStore();
 let cookieTokenStore: TokenStore | null = null;
 const cookieTokenStoreInitializer = (): TokenStore => {
   if (!isClient()) {
@@ -176,35 +185,6 @@ const cookieTokenStoreInitializer = (): TokenStore => {
 
   return cookieTokenStore;
 };
-
-const tokenStoreInitializers = new Map<TokenStoreOptions, () => TokenStore>([
-  ["cookie", cookieTokenStoreInitializer],
-  ["nextjs-cookie", () => {
-    if (isClient()) {
-      return cookieTokenStoreInitializer();
-    } else {
-      const store = new AsyncStore<TokenObject>();
-      store.set({
-        refreshToken: getCookie('stack-refresh'),
-        accessToken: getCookie('stack-access'),
-      });
-      store.onChange((value) => {
-        try {
-          setOrDeleteCookie('stack-refresh', value.refreshToken, { maxAge: 60 * 60 * 24 * 365 });
-          setOrDeleteCookie('stack-access', value.accessToken, { maxAge: 60 * 60 * 24 });
-        } catch (e) {
-          // ignore
-        }
-      });
-      return store;
-    }
-  }],
-  ["memory", () => memoryTokenStore],
-  [null, () => createEmptyTokenStore()],
-]);
-function getTokenStore(tokenStoreOptions: TokenStoreOptions) {
-  return (tokenStoreInitializers.get(tokenStoreOptions) ?? throwErr(`Invalid token store ${tokenStoreOptions}`))();
-}
 
 const loadingSentinel = Symbol("stackAppCacheLoadingSentinel");
 function useCache<D extends any[], T>(cache: AsyncCache<D, T>, dependencies: D, caller: string): T {
@@ -251,7 +231,7 @@ const createCacheByTokenStore = <D extends any[], T>(fetcher: (tokenStore: Token
       onSubscribe: ([tokenStore], refresh) => {
         // TODO find a *clean* way to not refresh when the token change was made inside the fetcher (for example due to expired access token)
         const handlerObj = tokenStore.onChange((newValue, oldValue) => {
-          if (JSON.stringify(newValue) === JSON.stringify(oldValue)) return;
+          if (newValue.refreshToken === oldValue?.refreshToken) return;
           refresh();
         });
         return () => handlerObj.unsubscribe();
@@ -262,9 +242,9 @@ const createCacheByTokenStore = <D extends any[], T>(fetcher: (tokenStore: Token
 
 
 class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends string = string> {
-  protected readonly _uniqueIdentifier: string;
+  protected _uniqueIdentifier: string | undefined = undefined;
   protected _interface: StackClientInterface;
-  protected readonly _tokenStoreOptions: TokenStoreOptions<HasTokenStore>;
+  protected readonly _tokenStoreInit: TokenStoreInit<HasTokenStore>;
   protected readonly _urlOptions: Partial<HandlerUrls>;
 
   private readonly __DEMO_ENABLE_SLIGHT_FETCH_DELAY = false;
@@ -292,7 +272,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     return await this._interface.listClientUserTeams(tokenStore);
   });
 
-  constructor(options:
+  constructor(protected readonly _options:
     & {
       uniqueIdentifier?: string,
       checkString?: string,
@@ -304,43 +284,105 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       }
     )
   ) {
-    if ("interface" in options) {
-      this._interface = options.interface;
+    if ("interface" in _options) {
+      this._interface = _options.interface;
     } else {
       this._interface = new StackClientInterface({
-        baseUrl: options.baseUrl ?? getDefaultBaseUrl(),
-        projectId: options.projectId ?? getDefaultProjectId(),
+        baseUrl: _options.baseUrl ?? getDefaultBaseUrl(),
+        projectId: _options.projectId ?? getDefaultProjectId(),
         clientVersion,
-        publishableClientKey: options.publishableClientKey ?? getDefaultPublishableClientKey(),
+        publishableClientKey: _options.publishableClientKey ?? getDefaultPublishableClientKey(),
       });
     }
 
-    this._tokenStoreOptions = options.tokenStore;
-    this._urlOptions = options.urls ?? {};
+    this._tokenStoreInit = _options.tokenStore;
+    this._urlOptions = _options.urls ?? {};
+  }
 
-    this._uniqueIdentifier = options.uniqueIdentifier ?? generateUuid();
-    if (allClientApps.has(this._uniqueIdentifier)) {
-      throw new StackAssertionError("A Stack client app with the same unique identifier already exists");
+  /**
+   * Cloudflare workers does not allow use of randomness on the global scope (on which the Stack app is probably
+   * initialized). For that reason, we generate the unique identifier lazily when it is first needed.
+   */
+  protected _getUniqueIdentifier() {
+    if (!this._uniqueIdentifier) {
+      this._uniqueIdentifier = this._options.uniqueIdentifier || generateUuid();
+      if (allClientApps.has(this._uniqueIdentifier)) {
+        throw new StackAssertionError("A Stack client app with the same unique identifier already exists");
+      }
+      allClientApps.set(this._uniqueIdentifier, [this._options.checkString ?? "default check string", this]);
     }
-    allClientApps.set(this._uniqueIdentifier, [options.checkString ?? "default check string", this]);
+    return this._uniqueIdentifier!;
   }
 
-  protected hasPersistentTokenStore(): this is StackClientApp<true, ProjectId> {
-    return this._tokenStoreOptions !== null;
-  }
+  private _memoryTokenStore = createEmptyTokenStore();
+  private _requestTokenStores = new Map<Request, TokenStore>();
+  protected _getTokenStore(overrideTokenStoreInit?: TokenStoreInit): TokenStore {
+    const tokenStoreInit = overrideTokenStoreInit === undefined ? this._tokenStoreInit : overrideTokenStoreInit;
 
-  protected _ensurePersistentTokenStore(): asserts this is StackClientApp<true, ProjectId>  {
-    if (!this.hasPersistentTokenStore()) {
-      throw new Error("Cannot call this function on a Stack app without a persistent token store. Make sure the tokenStore option is set to a non-null value when initializing Stack.");
+    switch (tokenStoreInit) {
+      case "cookie": {
+        return cookieTokenStoreInitializer();
+      }
+      case "nextjs-cookie": {
+        if (isClient()) {
+          return cookieTokenStoreInitializer();
+        } else {
+          const store = new AsyncStore<TokenObject>();
+          store.set({
+            refreshToken: getCookie('stack-refresh'),
+            accessToken: getCookie('stack-access'),
+          });
+          store.onChange((value) => {
+            try {
+              setOrDeleteCookie('stack-refresh', value.refreshToken, { maxAge: 60 * 60 * 24 * 365 });
+              setOrDeleteCookie('stack-access', value.accessToken, { maxAge: 60 * 60 * 24 });
+            } catch (e) {
+              // ignore
+            }
+          });
+          return store;
+        }
+      }
+      case "memory": {
+        return this._memoryTokenStore;
+      }
+      case null: {
+        return createEmptyTokenStore();
+      }
+      default: {
+        if (tokenStoreInit && typeof tokenStoreInit === "object" && "headers" in tokenStoreInit) {
+          if (this._requestTokenStores.has(tokenStoreInit)) return this._requestTokenStores.get(tokenStoreInit)!;
+          const cookieHeader = tokenStoreInit.headers.get("cookie");
+          const parsed = cookie.parse(cookieHeader || "");
+          const res = new AsyncStore<TokenObject>({
+            refreshToken: parsed['stack-refresh'] || null,
+            accessToken: parsed['stack-access'] || null,
+          });
+          this._requestTokenStores.set(tokenStoreInit, res);
+          return res;
+        }
+    
+        throw new Error(`Invalid token store ${tokenStoreInit}`);
+      }
     }
   }
 
-  protected isInternalProject(): this is { projectId: "internal" } {
+  protected _hasPersistentTokenStore(overrideTokenStoreInit?: TokenStoreInit): this is StackClientApp<true, ProjectId> {
+    return (overrideTokenStoreInit !== undefined ? overrideTokenStoreInit : this._tokenStoreInit) !== null;
+  }
+
+  protected _ensurePersistentTokenStore(overrideTokenStoreInit?: TokenStoreInit): asserts this is StackClientApp<true, ProjectId>  {
+    if (!this._hasPersistentTokenStore(overrideTokenStoreInit)) {
+      throw new Error("Cannot call this function on a Stack app without a persistent token store. Make sure the tokenStore option on the constructor is set to a non-null value when initializing Stack.\n\nStack uses token stores to access access tokens of the current user. For example, on web frontends it is commonly the string value 'cookies' for cookie storage.");
+    }
+  }
+
+  protected _isInternalProject(): this is { projectId: "internal" } {
     return this.projectId === "internal";
   }
 
   protected _ensureInternalProject(): asserts this is { projectId: "internal" } {
-    if (!this.isInternalProject()) {
+    if (!this._isInternalProject()) {
       throw new Error("Cannot call this function on a Stack app with a project ID other than 'internal'.");
     }
   }
@@ -412,24 +454,24 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
         });
       },
       async listTeams() {
-        const teams = await app._currentUserTeamsCache.getOrWait([getTokenStore(app._tokenStoreOptions)], "write-only");
+        const teams = await app._currentUserTeamsCache.getOrWait([app._getTokenStore()], "write-only");
         return teams.map((json) => app._teamFromJson(json));
       },
       useTeams() {
-        const teams = useCache(app._currentUserTeamsCache, [getTokenStore(app._tokenStoreOptions)], "user.useTeams()");
+        const teams = useCache(app._currentUserTeamsCache, [app._getTokenStore()], "user.useTeams()");
         return useMemo(() => teams.map((json) => app._teamFromJson(json)), [teams]);
       },
       onTeamsChange(callback: (value: Team[], oldValue: Team[] | undefined) => void) {
-        return app._currentUserTeamsCache.onChange([getTokenStore(app._tokenStoreOptions)], (value, oldValue) => {
+        return app._currentUserTeamsCache.onChange([app._getTokenStore()], (value, oldValue) => {
           callback(value.map((json) => app._teamFromJson(json)), oldValue?.map((json) => app._teamFromJson(json)));
         });
       },
       async listPermissions(scope: Team, options?: { direct?: boolean }): Promise<Permission[]> {
-        const permissions = await app._currentUserPermissionsCache.getOrWait([getTokenStore(app._tokenStoreOptions), scope.id, 'team', !!options?.direct], "write-only");
+        const permissions = await app._currentUserPermissionsCache.getOrWait([app._getTokenStore(), scope.id, 'team', !!options?.direct], "write-only");
         return permissions.map((json) => app._permissionFromJson(json));
       },
       usePermissions(scope: Team, options?: { direct?: boolean }): Permission[] {
-        const permissions = useCache(app._currentUserPermissionsCache, [getTokenStore(app._tokenStoreOptions), scope.id, 'team', !!options?.direct], "user.usePermissions()");
+        const permissions = useCache(app._currentUserPermissionsCache, [app._getTokenStore(), scope.id, 'team', !!options?.direct], "user.usePermissions()");
         return useMemo(() => permissions.map((json) => app._permissionFromJson(json)), [permissions]);
       },
       usePermission(scope: Team, permissionId: string): Permission | null {
@@ -468,6 +510,9 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     const currentUser: CurrentUser = {
       ...this._userFromJson(json),
       tokenStore,
+      async refreshAccessToken() {
+        await app._interface.refreshAccessToken(tokenStore);
+      },
       async updateSelectedTeam(team: Team | null) {
         await app._updateUser({ selectedTeamId: team?.id ?? null }, tokenStore);
       },
@@ -484,7 +529,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
         return app._updatePassword(options, tokenStore);
       },
     };
-    if (this.isInternalProject()) {
+    if (this._isInternalProject()) {
       const internalUser: CurrentInternalUser = {
         ...currentUser,
         createProject(newProject: ProjectUpdateOptions & { displayName: string }) {
@@ -552,6 +597,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       projectId: forProjectId,
       clientVersion,
       projectOwnerTokens: tokenStore,
+      refreshProjectOwnerTokens: async () => await this._interface.refreshAccessToken(tokenStore),
     });
   }
 
@@ -569,12 +615,16 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       throw new Error(`No URL for handler name ${handlerName}`);
     }
 
-    if (options?.replace) {
-      window.location.replace(url);
+    if (isReactServer) {
+      NextNavigation.redirect(url, options?.replace ? NextNavigation.RedirectType.replace : NextNavigation.RedirectType.push);
     } else {
-      window.location.assign(url);
+      if (options?.replace) {
+        window.location.replace(url);
+      } else {
+        window.location.assign(url);
+      }
+      await wait(2000);
     }
-    return await wait(2000);
   }
 
   async redirectToSignIn() { return await this._redirectTo("signIn"); }
@@ -620,15 +670,15 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
   async getUser(options: GetUserOptions & { or: 'throw' }): Promise<ProjectCurrentUser<ProjectId>>;
   async getUser(options?: GetUserOptions): Promise<ProjectCurrentUser<ProjectId> | null>;
   async getUser(options?: GetUserOptions): Promise<ProjectCurrentUser<ProjectId> | null> {
-    this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    this._ensurePersistentTokenStore(options?.tokenStore);
+    const tokenStore = this._getTokenStore(options?.tokenStore);
     const userJson = await this._currentUserCache.getOrWait([tokenStore], "write-only");
 
     if (userJson === null) {
       switch (options?.or) {
         case 'redirect': {
-          NextNavigation.redirect(this.urls.signIn, NextNavigation.RedirectType.replace);
-          throw new Error("redirect should never return!");
+          await this.redirectToSignIn();
+          break;
         }
         case 'throw': {
           throw new Error("User is not signed in but getUser was called with { or: 'throw' }");
@@ -646,16 +696,21 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
   useUser(options: GetUserOptions & { or: 'throw' }): ProjectCurrentUser<ProjectId>;
   useUser(options?: GetUserOptions): ProjectCurrentUser<ProjectId> | null;
   useUser(options?: GetUserOptions): ProjectCurrentUser<ProjectId> | null {
-    this._ensurePersistentTokenStore();
+    this._ensurePersistentTokenStore(options?.tokenStore);
 
     const router = NextNavigation.useRouter();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore(options?.tokenStore);
     const userJson = useCache(this._currentUserCache, [tokenStore], "useUser()");
 
     if (userJson === null) {
       switch (options?.or) {
         case 'redirect': {
-          router.replace(this.urls.signIn);
+          // Updating the router is not allowed during the component render function, so we do it in a different async tick
+          // The error would be: "Cannot update a component (`Router`) while rendering a different component."
+          runAsynchronously(async () => {
+            await wait(0);
+            router.replace(this.urls.signIn);
+          });
           suspend();
           throw new StackAssertionError("suspend should never return");
         }
@@ -676,7 +731,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   onUserChange(callback: (user: CurrentUser | null) => void) {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     return this._currentUserCache.onChange([tokenStore], (userJson) => {
       callback(this._currentUserFromJson(userJson, tokenStore));
     });
@@ -698,7 +753,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     password: string,
   }): Promise<KnownErrors["EmailPasswordMismatch"] | undefined> {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const errorCode = await this._interface.signInWithCredential(options.email, options.password, tokenStore);
     if (!errorCode) {
       await this.redirectToAfterSignIn({ replace: true });
@@ -711,7 +766,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     password: string,
   }): Promise<KnownErrors["UserEmailAlreadyExists"] | KnownErrors['PasswordRequirementsNotMet'] | undefined> {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const emailVerificationRedirectUrl = constructRedirectUrl(this.urls.emailVerification);
     const errorCode = await this._interface.signUpWithCredential(
       options.email, 
@@ -727,7 +782,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   async signInWithMagicLink(code: string): Promise<KnownErrors["MagicLinkError"] | undefined> {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const result = await this._interface.signInWithMagicLink(code, tokenStore);
     if (result instanceof KnownError) {
       return result;
@@ -741,7 +796,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   async callOAuthCallback() {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const result = await callOAuthCallback(this._interface, tokenStore, this.urls.oauthCallback);
     if (result) {
       if (result.newUser) {
@@ -793,7 +848,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   protected async _listOwnedProjects(): Promise<Project[]> {
     this._ensureInternalProject();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const json = await this._ownedProjectsCache.getOrWait([tokenStore], "write-only");
     return json.map((j) => this._projectAdminFromJson(
       j,
@@ -804,7 +859,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   protected _useOwnedProjects(): Project[] {
     this._ensureInternalProject();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const json = useCache(this._ownedProjectsCache, [tokenStore], "useOwnedProjects()");
     return useMemo(() => json.map((j) => this._projectAdminFromJson(
       j,
@@ -815,7 +870,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   protected _onOwnedProjectsChange(callback: (projects: Project[]) => void) {
     this._ensureInternalProject();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     return this._ownedProjectsCache.onChange([tokenStore], (projects) => {
       callback(projects.map((j) => this._projectAdminFromJson(
         j,
@@ -827,7 +882,7 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   protected async _createProject(newProject: ProjectUpdateOptions & { displayName: string }): Promise<Project> {
     this._ensureInternalProject();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const json = await this._interface.createProject(newProject, tokenStore);
     const res = this._projectAdminFromJson(
       json,
@@ -889,13 +944,13 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
           baseUrl: this._interface.options.baseUrl,
           projectId: this.projectId,
           publishableClientKey: this._interface.options.publishableClientKey,
-          tokenStore: this._tokenStoreOptions,
+          tokenStore: this._tokenStoreInit,
           urls: this._urlOptions,
-          uniqueIdentifier: this._uniqueIdentifier,
+          uniqueIdentifier: this._getUniqueIdentifier(),
         };
       },
       setCurrentUser: (userJsonPromise: Promise<UserJson | null>) => {
-        runAsynchronously(this._currentUserCache.forceSetCachedValueAsync([getTokenStore(this._tokenStoreOptions)], userJsonPromise));
+        runAsynchronously(this._currentUserCache.forceSetCachedValueAsync([this._getTokenStore()], userJsonPromise));
       },
     };
   };
@@ -941,7 +996,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     | StackServerAppConstructorOptions<HasTokenStore, ProjectId>
     | {
       interface: StackServerInterface,
-      tokenStore: TokenStoreOptions<HasTokenStore>,
+      tokenStore: TokenStoreInit<HasTokenStore>,
       urls: Partial<HandlerUrls> | undefined,
     }
   ) {
@@ -1013,15 +1068,15 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
         });
       },
       async listTeams() {
-        const teams = await app._serverTeamsCache.getOrWait([getTokenStore(app._tokenStoreOptions)], "write-only");
+        const teams = await app._serverTeamsCache.getOrWait([app._getTokenStore()], "write-only");
         return teams.map((json) => app._serverTeamFromJson(json));
       },
       useTeams() {
-        const teams = useCache(app._serverTeamsCache, [getTokenStore(app._tokenStoreOptions)], "user.useTeams()");
+        const teams = useCache(app._serverTeamsCache, [app._getTokenStore()], "user.useTeams()");
         return useMemo(() => teams.map((json) => app._serverTeamFromJson(json)), [teams]);
       },
       onTeamsChange(callback: (value: ServerTeam[], oldValue: ServerTeam[] | undefined) => void) {
-        return app._serverTeamsCache.onChange([getTokenStore(app._tokenStoreOptions)], (value, oldValue) => {
+        return app._serverTeamsCache.onChange([app._getTokenStore()], (value, oldValue) => {
           callback(value.map((json) => app._serverTeamFromJson(json)), oldValue?.map((json) => app._serverTeamFromJson(json)));
         });
       },
@@ -1060,6 +1115,9 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
     const currentUser: CurrentServerUser = {
       ...nonCurrentServerUser,
       tokenStore,
+      async refreshAccessToken() {
+        await app._interface.refreshAccessToken(tokenStore);
+      },
       async delete() {
         const res = await nonCurrentServerUser.delete();
         await app._refreshUser(tokenStore);
@@ -1087,7 +1145,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
       },
     };
 
-    if (this.isInternalProject()) {
+    if (this._isInternalProject()) {
       const internalUser: CurrentInternalServerUser = {
         ...currentUser,
         createProject(newProject: ProjectUpdateOptions & { displayName: string }) {
@@ -1169,7 +1227,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   async getServerUser(): Promise<ProjectCurrentSeverUser<ProjectId> | null> {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const userJson = await this._currentServerUserCache.getOrWait([tokenStore], "write-only");
     return this._currentServerUserFromJson(userJson, tokenStore);
   }
@@ -1182,7 +1240,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
   useServerUser(options?: { required: boolean }): ProjectCurrentSeverUser<ProjectId> | null {
     this._ensurePersistentTokenStore();
 
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     const userJson = useCache(this._currentServerUserCache, [tokenStore], "useServerUser()");
 
     return useMemo(() => {
@@ -1196,7 +1254,7 @@ class _StackServerAppImpl<HasTokenStore extends boolean, ProjectId extends strin
 
   onServerUserChange(callback: (user: CurrentServerUser | null) => void) {
     this._ensurePersistentTokenStore();
-    const tokenStore = getTokenStore(this._tokenStoreOptions);
+    const tokenStore = this._getTokenStore();
     return this._currentServerUserCache.onChange([tokenStore], (userJson) => {
       callback(this._currentServerUserFromJson(userJson, tokenStore));
     });
@@ -1335,6 +1393,7 @@ class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string
         clientVersion,
         ..."projectOwnerTokens" in options ? {
           projectOwnerTokens: options.projectOwnerTokens,
+          refreshProjectOwnerTokens: options.refreshProjectOwnerTokens,
         } : {
           publishableClientKey: options.publishableClientKey ?? getDefaultPublishableClientKey(),
           secretServerKey: options.secretServerKey ?? getDefaultSecretServerKey(),
@@ -1457,7 +1516,8 @@ type RedirectToOptions = {
 };
 
 type Auth<T, C> = {
-  readonly tokenStore: ReadonlyTokenStore,
+  readonly tokenStore: TokenStore,
+  refreshAccessToken(this: T): Promise<void>,
   updateSelectedTeam(this: T, team: Team | null): Promise<void>,
   update(this: T, user: C): Promise<void>,
   signOut(this: T): Promise<void>,
@@ -1660,6 +1720,7 @@ export type OAuthProviderConfig = OAuthProviderConfigJson;
 
 export type GetUserOptions = {
   or?: 'redirect' | 'throw' | 'return-null',
+  tokenStore?: TokenStoreInit,
 };
 
 type SplitArgs<T extends any[], U extends number> = [