+++ apps/custom-pages-example/CHANGELOG.md
@@ -1,5 +1,28 @@
 # demo-app
 
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.4.11
+
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.4.8
+  - @stackframe/stack@2.4.10
+
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.4.7
+  - @stackframe/stack@2.4.9
+
 ## 1.1.8
 
 ### Patch Changes