+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-shared
 
+## 2.4.8
+
+### Patch Changes
+
+- Improved UUID generation
+
+## 2.4.7
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-sc@1.5.3
+
 ## 2.4.6
 
 ### Patch Changes